*) Code from I50, Lohx and Me - Added: Untested support for all non RR5 RAs apart from 3. *) Code from Mannen - Changed: The LootGeneratorRandom to drop only your the zones realm items when on a rvr server. *) Code from Tiv - Bugfix: Removed alot of unnessary packets from keep reset - Bugfix: Fixed the level and target level display on gamekeepcomponent interact. *) Code from Me - Bugfix: Setting a controlled NPC to passive will now clear it's aggro list (Thanks sirru) - Bugfix: Crafting speed server property bonus should now work. (Thanks Stigmatic) - Added: GameLivingEvent.CastFinished - Bugfix: The PvE and PvP damage server property boni should now work with controlled NPCs too. - Changed: Bindpoints are now loaded into memory and then checked instead of performing a db query each time the /bind command is typed. - Bugfix: Fixed the players in area check for keep claims (Thanks Tiv) - Changed: All GameKeepComponent and GameKeepDoor will repair at 5% every 30 minutes, unless someone knows they behave differently (see forums) - Bugfix: Moving too far away from your trade window will now close it. - Bugfix: Realm Ability bonuses will now display in the character window for 1.75 clients. (Thanks Tiv) - Added: New Server Properties (load_examples, disabled_regions, disable_tutorial, allow_all_realms, disable_catacombs_classes) they are all pretty self explanatory. - Bugfix: Players can now trade with GMs from other realms.

+++ GameServerScripts/playercommands/guild.cs
@@ -1314,10 +1314,10 @@ namespace DOL.GS.Scripts
 								client.Out.SendMessage("You must specify a level to target for upgrade.", eChatType.CT_System, eChatLoc.CL_SystemWindow);
 								return 1;
 							}
-							int targetlevel = 0;
+							byte targetlevel = 0;
 							try
 							{
-								targetlevel = Convert.ToInt32(args[2]);
+								targetlevel = Convert.ToByte(args[2]);
 								if (targetlevel > 10 || targetlevel < 1)
 									return 0;
 							}
